requirements.txt
bidict==0.21.3
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.1
Flask-SocketIO==4.3.1
idna==3.2
imagezmq==1.1.1
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
python-engineio==3.13.2
python-socketio==4.6.0
pyzmq==22.3.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
websocket-client==1.2.1
Werkzeug==2.0.1
